requirements.txt
streamlit==1.35.0
requests==2.31.0
google-generativeai
python-dotenv

